Queue Length tracker and discarded Jobs tracker

+++ testbench/src/main/java/org/com/autoscaler/queue/IQueue.java
@@ -30,6 +30,6 @@ public interface IQueue {
 
     int currentLevelInTasks();
 
-    long currentLevelInPercent();
+    double currentLevelInPercent();
 
 }
